requirements.txt
google-api-python-client==2.86.0
google-auth==2.21.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
reportlab==3.6.13